integrated_vmae/integrated_mala_core/requirements.txt
# MALA项目依赖列表
# ==================
# 使用 pip install -r requirements.txt 安装所有依赖

# 深度学习框架
torch>=1.10.0
torchvision>=0.11.0

# 数值计算
numpy>=1.19.0

# 图像处理
opencv-python>=4.5.0
Pillow>=8.0.0

# 科学计算
scikit-image>=0.18.0
scipy>=1.6.0
pandas>=1.2.0

# 可视化
matplotlib>=3.3.0
seaborn>=0.11.0

# 深度学习辅助
einops>=0.4.0

# 其他工具
tqdm>=4.60.0
pyyaml>=5.4.0

# 可选：LaMa图像修复（需要单独安装）
# simple-lama-inpainting>=0.1.0

# 可选：Jupyter支持（如果需要在Jupyter中运行）
# ipykernel>=6.0.0
# jupyter>=1.0.0
